Since this version HeroApi provides through DI. Interface HeroApiService is renamed to HeroApi
