+++ app/src/main/java/com/example/marvel_app_project/data/HeroDatabase.kt
@@ -1,8 +1,6 @@
 package com.example.marvel_app_project.data
 
-import android.content.Context
 import androidx.room.Database
-import androidx.room.Room
 import androidx.room.RoomDatabase
 import com.example.marvel_app_project.models.data.HeroEntity
 
@@ -14,23 +12,23 @@ import com.example.marvel_app_project.models.data.HeroEntity
 abstract class HeroDatabase: RoomDatabase() {
     abstract fun heroDao(): HeroDao
 
-    companion object{
-        @Volatile
-        private var INSTANCE: HeroDatabase? = null
-
-        fun getDatabase(context: Context): HeroDatabase {
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    HeroDatabase::class.java,
-                    "hero.db"
-                ).build()
-                INSTANCE = instance
-                instance
-            }
-
-        }
-    }
+//    companion object{
+//        @Volatile
+//        private var INSTANCE: HeroDatabase? = null
+//
+//        fun getDatabase(context: Context): HeroDatabase {
+//            return INSTANCE ?: synchronized(this) {
+//                val instance = Room.databaseBuilder(
+//                    context.applicationContext,
+//                    HeroDatabase::class.java,
+//                    "hero.db"
+//                ).build()
+//                INSTANCE = instance
+//                instance
+//            }
+//
+//        }
+//    }
 
 
 }
